+++ include/chemfiles.hpp
@@ -19,15 +19,12 @@
 #include "chemfiles/config.hpp"
 
 #include "chemfiles/Error.hpp"
-#include "chemfiles/Logger.hpp"
-
 #include "chemfiles/Atom.hpp"
 #include "chemfiles/Frame.hpp"
 #include "chemfiles/Topology.hpp"
 #include "chemfiles/Residue.hpp"
 #include "chemfiles/Trajectory.hpp"
 #include "chemfiles/UnitCell.hpp"
-
 #include "chemfiles/Selections.hpp"
 
 #endif // CHEMFILES_HPP