+++ include/chemfiles/utils.hpp
@@ -8,8 +8,6 @@
 #ifndef CHEMFILES_UTILS_HPP
 #define CHEMFILES_UTILS_HPP
 
-#include "chemfiles/Logger.hpp"
-
 #include <vector>
 #include <string>
 #include <sstream>
@@ -47,9 +45,8 @@ inline std::string trim(const std::string& str) {
 #elif defined(_MSC_VER)
     #define unreachable() __assume(false)
 #else
-    #define unreachable() do {                                                     \
-        Logger::error("Entered unreachable code at ", __FILE__, ":", __LINE__);    \
-        abort();                                                                   \
+    #define unreachable() do {                     \
+        throw Error("Entered unreachable code");  \
     } while (false)
 #endif